NF: Plain condor-based staging of job input files
Bits from `run` are used to expand the inputs, `status()` is used
to obtain a list of all matching files, regardless of whether they
are just under Git or annex, and condor_chirp moves them from the
submit machine to the execute node.

Much room for tuning, but the basic functionality works.